Moved NativePointerInput to address circular dependency
